backend/requirements.txt
# Backend Python dependencies for jobMCP-deepcrawl
# Note: Some tools (pytesseract, pdf2image) require system binaries:
# - Tesseract OCR (https://github.com/tesseract-ocr/tesseract)
# - Poppler for PDF rendering (for pdf2image)
# Install those separately on your OS.

# Web framework & ASGI server
fastapi>=0.110
uvicorn>=0.23
starlette>=0.36
python-multipart>=0.0.9

# HTTP client
httpx>=0.27

# Config & validation
pydantic
pydantic-settings

# Database & Auth
motor>=3.3.1
pymongo>=4.6.0
PyJWT>=2.8.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4

# MCP protocol & helpers
fastmcp>=0.1.8
mcp>=0.2.0

# Crawling, parsing, and search
crawl4ai>=0.3.3
beautifulsoup4>=4.12
duckduckgo-search>=5.3.0

# CV OCR pipeline
pdfplumber>=0.11.0
pytesseract>=0.3.10
pdf2image>=1.17.0
numpy>=1.26.0
scikit-learn>=1.4.0
faiss-cpu>=1.7.4
langdetect>=1.0.9
sentence-transformers>=2.2.2

# System & utilities
psutil>=5.9.0

# Optional: if running on Windows and encountering SSL/cert issues
# certifi>=2024.2.2

# Hugging Face Transformers (Phi-3 Mini host)
transformers>=4.42.0
safetensors>=0.4.2
accelerate>=0.30.0
# NOTE: Install torch separately with the right build for your GPU/CPU. See commands in README or assistant message.
